stabilized backend

+++ src/test/java/com/ao/portfolio/service/PositionServiceTest.java
@@ -1,10 +1,11 @@
 package com.ao.portfolio.service;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.mock;
+import java.math.BigDecimal;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import static org.mockito.Mockito.mock;
 
 import com.ao.portfolio.entity.Position;
 import com.ao.portfolio.repository.PositionRepository;
@@ -24,7 +25,7 @@ public class PositionServiceTest {
         Position position = new Position();
         position.setSymbol("AAPL");
         position.setQuantity(10);
-        position.setAvgPrice(150);
+        position.setAvgPrice(BigDecimal.valueOf(150));
 
         double result = positionService.calculatePnL(position, 160);
 
@@ -36,7 +37,7 @@ public class PositionServiceTest {
         Position position = new Position();
         position.setSymbol("AAPL");
         position.setQuantity(10);
-        position.setAvgPrice(150);
+        position.setAvgPrice(BigDecimal.valueOf(150));
 
         double result = positionService.calculatePnL(position, 140);
